Move RealTestScenarioSeeder into :test-seeder-internal
Creates a new -internal module that contains RealTestScenarioSeeder and
the TestScenario enum, wired into :app via internalImplementation so the
real impl ships only with internal builds. Play and F-Droid flavours get
a NoOpTestScenarioSeeder so LaunchViewModel's seeding call hits a
guaranteed no-op there. RealOmnibarPositionWriter moves to the internal
flavour source set in :app, keeping SettingsDataStore writes out of
release builds.
