+++ test-seeder/test-seeder-internal/src/test/java/com/duckduckgo/testseeder/internal/RealTestScenarioSeederTest.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.app.launch
+package com.duckduckgo.testseeder.internal
 
 import com.duckduckgo.common.test.CoroutineTestRule
 import com.duckduckgo.duckchat.api.DuckChat
